resrc: Move the resource request code into its own module
Breaks the close association of a resource request to a resource tree.
Will allow a resource request to include other graph associations
(e.g., flow) independently from its support for a resource tree.
This commit introduces no functional changes to the code.
