+++ sched/schedplugin1.c
@@ -44,6 +44,7 @@
 #include "src/common/libutil/xzmalloc.h"
 #include "resrc.h"
 #include "resrc_tree.h"
+#include "resrc_reqst.h"
 #include "schedsrv.h"
 
 //TODO: this plugin inherently must know the inner structure of the opaque
